libmachine: split swarm and engine options
